+++ src/core/client/systems/nametag.ts
@@ -4,6 +4,7 @@ import { SHARED_CONFIG } from '../../shared/configurations/shared';
 import { SYSTEM_EVENTS } from '../../shared/enums/system';
 import { distance2d } from '../../shared/utility/vector';
 import { drawText2D } from '../utility/text';
+import { Timer } from '../utility/timers';
 
 const drawDistance = 50;
 let interval;
@@ -12,7 +13,7 @@ let shouldScreenShake = false;
 alt.onServer(SYSTEM_EVENTS.TICKS_START, handleStart);
 
 function handleStart() {
-    interval = alt.setInterval(drawNametags, 0);
+    interval = Timer.createInterval(drawNametags, 0, 'nametag.ts');
     shouldScreenShake = SHARED_CONFIG.FORCE_SCREEN_SHAKE;
 }