+++ src/core/client/systems/tick.ts
@@ -1,12 +1,13 @@
 import * as alt from 'alt-client';
 import { SYSTEM_EVENTS } from '../../shared/enums/system';
+import { Timer } from '../utility/timers';
 
 const pingEvery = 5000;
 
 alt.onServer(SYSTEM_EVENTS.TICKS_START, startTick);
 
 function startTick() {
-    alt.setInterval(handlePing, pingEvery);
+    Timer.createInterval(handlePing, pingEvery, 'tick.ts');
 }
 
 /**